% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls-package.R
\docType{package}
\name{lls-package}
\alias{lls-package}
\title{lls: Local Least Squares Estimation for Information Provision Experiments}
\description{
Implements the Local Least Squares (LLS) estimator to recover Average Partial
Effects (APE) in information provision experiments, robust to dependence between
belief updating and belief effects. Provides plotting helpers and example data.
}
\seealso{
\link{lls}, \link{plot.lls}
}
\author{
\strong{Maintainer}: Dylan Balla-Elliott \email{dballaelliott@gmail.com}

}
\keyword{package}
