% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls.R
\name{print.lls}
\alias{print.lls}
\alias{summary.lls}
\title{Print method for lls objects}
\usage{
\method{print}{lls}(lls.return, ...)

\method{summary}{lls}(lls.return, ...)
}
\arguments{
\item{lls.return}{An lls object}

\item{...}{Additional arguments (ignored)}

\item{object}{An lls object}
}
\description{
Print method for lls objects

Summary method for lls objects
   just an alias for print
}
